[4909] Move CreateForkedStudioMutationRunner to sirius-web-tests
Bug: https://github.com/eclipse-sirius/sirius-web/issues/4909
