+++ packages/sirius-web/backend/sirius-web-tests/pom.xml
@@ -90,6 +90,11 @@
 			<artifactId>sirius-components-portals-tests</artifactId>
 			<version>2025.4.1</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-web-view-fork</artifactId>
+			<version>2025.4.1</version>
+		</dependency>
 	</dependencies>
 
 	<build>
