+++ packages/sirius-web/backend/sirius-web/src/test/java/org/eclipse/sirius/web/application/controllers/studiofork/TestCreateForkedStudio.java
@@ -30,18 +30,17 @@ import org.eclipse.sirius.components.collaborative.tables.TableRefreshedEventPay
 import org.eclipse.sirius.components.core.api.IURLParser;
 import org.eclipse.sirius.components.view.emf.IRepresentationDescriptionIdProvider;
 import org.eclipse.sirius.web.AbstractIntegrationTests;
-import org.eclipse.sirius.web.application.controllers.studiofork.graphql.CreateForkedStudioMutationRuner;
 import org.eclipse.sirius.web.application.project.dto.CreateProjectSuccessPayload;
 import org.eclipse.sirius.web.data.StudioIdentifiers;
 import org.eclipse.sirius.web.domain.boundedcontexts.project.services.api.IProjectSearchService;
 import org.eclipse.sirius.web.domain.boundedcontexts.projectsemanticdata.services.api.IProjectSemanticDataSearchService;
 import org.eclipse.sirius.web.domain.boundedcontexts.representationdata.services.api.IRepresentationMetadataSearchService;
 import org.eclipse.sirius.web.domain.boundedcontexts.semanticdata.repositories.ISemanticDataRepository;
-import org.eclipse.sirius.web.domain.boundedcontexts.semanticdata.services.api.ISemanticDataSearchService;
-import org.eclipse.sirius.web.view.fork.dto.CreateForkedStudioInput;
 import org.eclipse.sirius.web.tests.data.GivenSiriusWebServer;
 import org.eclipse.sirius.web.tests.services.api.IGivenCreatedTableSubscription;
 import org.eclipse.sirius.web.tests.services.api.IGivenInitialServerState;
+import org.eclipse.sirius.web.tests.services.fork.CreateForkedStudioMutationRunner;
+import org.eclipse.sirius.web.view.fork.dto.CreateForkedStudioInput;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -70,16 +69,13 @@ public class TestCreateForkedStudio extends AbstractIntegrationTests {
     private IGivenCreatedTableSubscription givenCreatedTableSubscription;
 
     @Autowired
-    private CreateForkedStudioMutationRuner createForkedStudioMutationRuner;
+    private CreateForkedStudioMutationRunner createForkedStudioMutationRunner;
 
     @Autowired
     private IProjectSearchService projectSearchService;
 
     @Autowired
     private IRepresentationMetadataSearchService representationMetadataSearchService;
-
-    @Autowired
-    private ISemanticDataSearchService semanticDataSearchService;
 
     @Autowired
     private ISemanticDataRepository semanticDataRepository;
@@ -139,7 +135,7 @@ public class TestCreateForkedStudio extends AbstractIntegrationTests {
 
         Runnable forkStudio = () -> {
             var input = new CreateForkedStudioInput(UUID.randomUUID(), StudioIdentifiers.INSTANCE_EDITING_CONTEXT_ID, representationId.get(), "");
-            var result = this.createForkedStudioMutationRuner.run(input);
+            var result = this.createForkedStudioMutationRunner.run(input);
 
             String typename = JsonPath.read(result, "$.data.createForkedStudio.__typename");
             assertThat(typename).isEqualTo(CreateProjectSuccessPayload.class.getSimpleName());
